try different themes arch
