+++ packages/nuxt/playground/app/app.config.ts
@@ -3,6 +3,8 @@ export default defineAppConfig({
     colors: {
       primary: 'blue',
     },
-    components: {}
+    components: {
+      button: {}
+    }
   }
 })
