--- packages/themes/src/components/prose/a.ts
@@ -1,11 +1,0 @@
-import { tv } from 'tailwind-variants'
-
-export const proseA = tv({
-  base: 'text-[var(--ui-primary)] border-b border-transparent hover:border-[var(--ui-primary)] font-medium focus-visible:outline-[var(--ui-primary)] [&>code]:border-dashed hover:[&>code]:border-[var(--ui-primary)] hover:[&>code]:text-[var(--ui-primary)]',
-})
-
-export type ProseAProps = {
-  href?: string
-  target?: '_blank' | '_parent' | '_self' | '_top' | (string & object) | null | undefined
-  class?: any
-}
