+++ packages/themes/src/components/prose/codeGroup.ts
@@ -11,8 +11,3 @@ export const proseCodeGroup = tv({
   }
 })
 
-export type CodeGroupProps = {
-  defaultValue?: string
-  class?: any
-  sync?: string
-}
