+++ packages/themes/src/components/prose/li.ts
@@ -3,7 +3,3 @@ import { tv } from 'tailwind-variants'
 export const proseLi = tv({
   base: 'my-1.5 ps-1.5 leading-7 [&>ul]:my-0'
 })
-
-export type ProseLiProps = {
-  class?: any
-}
