+++ packages/themes/src/components/prose/ol.ts
@@ -4,6 +4,3 @@ export const proseOl = tv({
   base: 'list-decimal pl-6 my-5'
 })
 
-export type ProseOlProps = {
-  class?: any
-}
