--- packages/themes/src/components/prose/p.ts
@@ -1,9 +1,0 @@
-import { tv } from 'tailwind-variants'
-
-export const proseP = tv({
-  base: 'my-5 leading-7'
-})
-
-export type ProsePProps = {
-  class?: any
-}
