+++ packages/themes/src/components/prose/pre.ts
@@ -18,14 +18,3 @@ export const prosePre = tv({
     }
   }
 })
-
-export type ProsePreProps = {
-  icon?: string
-  code?: string
-  language?: string
-  filename?: string
-  highlights?: number[]
-  hideHeader?: boolean
-  meta?: string
-  class?: any
-}
