+++ packages/themes/src/index.ts
@@ -1,6 +1,11 @@
 import * as components from './components'
+import * as types from './types'
 
 export * from './components'
+export * from './types'
+
+export { components }
+export { types }
 
 /**
  * Type helper to extract slot names from a component
@@ -13,6 +18,9 @@ type ComponentConfig<T> = {
     : never
 } | string
 
+/**
+ * Every component theme
+ */
 export type Themes = {
   [K in keyof typeof components]?: ComponentConfig<(typeof components)[K]>
 }
