+++ packages/themes/src/types/index.ts
@@ -2,11 +2,11 @@
 export * from './button'
 
 // Prose
+export * from './prose/callout'
+export * from './prose/code'
+export * from './prose/codeGroup'
 export * from './prose/li'
 export * from './prose/ul'
 export * from './prose/ol'
-export * from './prose/code'
 export * from './prose/pre'
-export * from './prose/codeGroup'
-export * from './prose/callout'
 export * from './prose/typography'
