+++ packages/themes/src/types/prose/codeGroup.ts
@@ -1,0 +1,5 @@
+export type CodeGroupProps = {
+  defaultValue?: string
+  class?: any
+  sync?: string
+}
