+++ packages/themes/src/types/prose/li.ts
@@ -1,0 +1,3 @@
+export type ProseLiProps = {
+  class?: any
+}
