+++ packages/themes/src/types/prose/ol.ts
@@ -1,0 +1,3 @@
+export type ProseOlProps = {
+  class?: any
+}
