+++ packages/themes/src/types/prose/pre.ts
@@ -1,0 +1,10 @@
+export type ProsePreProps = {
+  icon?: string
+  code?: string
+  language?: string
+  filename?: string
+  highlights?: number[]
+  hideHeader?: boolean
+  meta?: string
+  class?: any
+}
