+++ packages/themes/src/types/prose/typography.ts
@@ -1,0 +1,29 @@
+export type ProseH1Props = {
+  id?: string
+  class?: any
+}
+
+export type ProseH2Props = {
+  id?: string
+  class?: any
+}
+
+export type ProseH3Props = {
+  id?: string
+  class?: any
+}
+
+export type ProseH4Props = {
+  id?: string
+  class?: any
+}
+
+export type ProsePProps = {
+  class?: any
+}
+
+export type ProseAProps = {
+  href?: string
+  target?: '_blank' | '_parent' | '_self' | '_top' | (string & object) | null | undefined
+  class?: any
+}
