+++ packages/themes/src/types/prose/ul.ts
@@ -1,0 +1,3 @@
+export type ProseUlProps = {
+  class?: any
+}
